Add support for completions without `textEdit`
The NodeJS LSP does not always provide a `textEdit` when returning completion
results: if completion happens at the end of a word, it just responds with an
`insertText`, but no text editing spec.

According to
[the 3.17 LSP spec](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_completion),
in such a case the editor is supposed to delete the current word and replace it
with the insert text, after filtering candidates from the result set.

Since Acme does not have explicit language support (and we probably don't want
that in `acme-lsp` either), just do the equivalent of ^W and be done.

Use Levensthein distance for filtering completion candidates.

This is not yet good enough:

- it does too many things with the window, probably (like too many `dot=addr`, `addr=dot` control messages)
	- this also means that when not inserting a result (i.e. running plain `L comp`, or because there's no match), the current word stays marked
- there should be some sort of fuzzy matching going on, similar to what gopls does internally:
	https://github.com/golang/tools/blob/master/gopls/internal/fuzzy/matcher.go#L14